requirements.txt
matplotlib
numpy
pandas
scikit-learn
tensorflow-gpu
tensorflow-datasets
ipykernel
tqdm
wandb
einops
jax=0.4.1
jaxlib=0.4.1
flax=0.5.1